ecommerce/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
gunicorn
psycopg2-binary
celery==5.3.6
redis==5.0.3
python-dotenv==1.0.1
